// src/generated/schemas/BY.ts
/*!
 * Generated by addressfield. Do not edit.
 *
 * Google libaddressinput metadata
 * https://github.com/google/libaddressinput
 *
 * Creative Commons Attribution 4.0 International (CC BY 4.0)
 * https://creativecommons.org/licenses/by/4.0/
 *
 * Changes: defaults applied, fields renamed, data split into modules, Effect schemas added.
 */
import * as Schema from "effect/Schema";
import { BY } from "../countries/BY.ts";
import { text, addressLines, postalCode } from "../../schema-fields.ts";

export const validationCoverage = "country" as const;
export const BYAddressSchema = Schema.Struct({
  countryCode: Schema.Literal(BY.countryCode),
  recipient: Schema.optionalKey(text),
  organization: Schema.optionalKey(text),
  addressLines: addressLines,
  dependentLocality: Schema.optionalKey(text),
  locality: text,
  administrativeArea: Schema.optionalKey(text),
  postalCode: Schema.optionalKey(postalCode(BY.postalCodePattern)),
  sortingCode: Schema.optionalKey(text),
});