// src/generated/schemas/DE.ts
/*!
 * Generated by addressfield-ts. Do not edit.
 *
 * Google libaddressinput metadata
 * https://github.com/google/libaddressinput
 *
 * Creative Commons Attribution 4.0 International (CC BY 4.0)
 * https://creativecommons.org/licenses/by/4.0/
 */
import * as Schema from "effect/Schema";
import { AddressLines, AddressText, addressChecks, postalCode } from "../../validation.ts";
import { DE } from "../countries/DE.ts";
import { records } from "../regions/DE.ts";

export const DEAddressSchema = Schema.Struct({
  countryCode: Schema.Literal(DE.countryCode),
  recipient: Schema.optionalKey(AddressText),
  organization: Schema.optionalKey(AddressText),
  addressLines: AddressLines,
  dependentLocality: Schema.optionalKey(AddressText),
  locality: AddressText,
  administrativeArea: Schema.optionalKey(AddressText),
  postalCode: postalCode(DE),
  sortingCode: Schema.optionalKey(AddressText),
})
  .check(addressChecks(DE, records))
  .annotate({ identifier: "DEAddress" });