// src/generated/schemas/US.ts
/*!
 * Generated by addressfield-ts. Do not edit.
 *
 * Google libaddressinput metadata
 * https://github.com/google/libaddressinput
 *
 * Creative Commons Attribution 4.0 International (CC BY 4.0)
 * https://creativecommons.org/licenses/by/4.0/
 */
import * as Schema from "effect/Schema";
import { AddressLines, AddressText, addressChecks, postalCode } from "../../validation.ts";
import { US } from "../countries/US.ts";
import { records } from "../regions/US.ts";

export const USAddressSchema = Schema.Struct({
  countryCode: Schema.Literal(US.countryCode),
  recipient: Schema.optionalKey(AddressText),
  organization: Schema.optionalKey(AddressText),
  addressLines: AddressLines,
  dependentLocality: Schema.optionalKey(AddressText),
  locality: AddressText,
  administrativeArea: AddressText,
  postalCode: postalCode(US),
  sortingCode: Schema.optionalKey(AddressText),
})
  .check(addressChecks(US, records))
  .annotate({ identifier: "USAddress" });